Add files via upload
Index based method functions have been updated.   
Added Liz's stable period function that is an extension of AIM.  We will have to figure out the best way to utilize the function as I simply included it, but now AIM is in the code twice.  

The catch curve and expanded biomass functions have been modified so that they can each take different proxy Fmsy values 1. spr, 2 stable historical period, 3 natural mortality.  They use the Baranov catch eq for 1 and 3 to derive catch advise. 
Expanded biomass is back to one function and I removed the larger grouping list as it is not needed.   We agreed to use the true value of q for the expanded biomass method, but would use the historical q if there is a q ramp.  
I also pulled the SPR function up to the top and run it independent of the index based methods and the output is simply added to the y list.  I believe this should work and might shave the tiniest amounts off of run time.