got rid of stable period in M_CC,
I added this and then immediately got rid of it because I don't think it makes sense.  This approach would use expanded survey biomass to identify exploitation rate, but a different biomass to calculate catch advice.  Thus, relying on two different "estimates" of stock biomass, which seemed odd to me.