Add files via upload
Added Joe Langan's dynamic linear model function JoeDLM=function(y) to wham_mse_functions.R .  The function requires two packages 'dlm' and 'RandomFieldsUtils'.  Currently there are in the function, but not sure if there is a better location for loading all the required packages.  The function uses dynamic linear modeling to forecast the catch within specifications.  The function can forecast as many years of catch as needed to cover the time between assessments.  Currently the default is two, but this should be controlled at a master level and not just for this function.  This function can take up to sixty seconds to run and has multiple fitting routines within it that could crash if catch went to zero or something.