Add files via upload
Added Joe Langan's dynamic linear model function JoeDLM=function(y) to wham_mse_functions.R .  The function requires two packages 'dlm' and 'RandomFieldsUtils'.  Currently there are in the function, but not sure if there is a better location for loading all the required packages.  The function uses dynamic linear modeling to forecast the catch within specifications.  The function can forecast as many years of catch as needed to cover the time between assessments.  Currently the default is two, but this should be controlled at a master level and not just for this function.  The function takes 20-60 seconds to run

Updated the survey selectivity in the SPR function to use the true first year selectivity.  If there is a change in selectivity, it will be mis-specified as we discussed.  SPR is the only function using survey selectivity and it is currently hard coded. within the SPR function.