Update wham_mse_functions.R
Modified call to input.IBM function to accept input$adv.yr, the number of years between assessments.  Modified JoeDLM function to accept this value to object n.ahead. and modified skate_CR to output the smooth index and catch as diagnostics along with the catch advise as a list.  updated ensemble to only take the first object in the output list from skate_CR