Update wham_mse_functions.R
Adjusted the indexing to provide the correct years of spring survey, fall survey, annual survey and catch to the IBM functions.  As discussed on the call, if the assessment is done mid year in year 50 to provide catch advise for year 51, the available data are: spring survey through year 50, fall survey through year 49, and catch through year 49.  

The annual survey is the mean of the spring survey in year t and the fall survey in year t-1 to provide an estimate of abundance for Jan 1, in year t.  This means that in year one, there is a spring survey, but no fall survey the year prior so the mean is roughly half the true value (spr +0)/2.  The annual survey index has been adjusted to leave out year one.  The annual survey index (y$index) goes from year 2:t.  Indexing the annual survey this way means the survey and catch vectors are the same length.  For IBM functions that needed survey and catch to be the same length, they should be all set.  It does mean that the catch that occurs in year one is now aligned with the estimated survey biomass on Jan 1, year two.

The seasonal survey index goes from year 1 to year t so year one is included.  The fall survey in the terminal year is now an NA however as this value is not available to the assessment.  I made it an NA instead of a zero so that functions would produce an error if it is not accounted for.  The same is true for the numbers at age.  Years 1:t are available, but the terminal year in the fall is a row of NA's.  I have not run it through all the functions yet, but hopefully they do not all blow up.