Update IBMWG_Manuscript.Rmd
Thank you Team Scrapple for pulling this together.  

It is a huge amount of information to pull together.  Overall I think it looks great.  I made a few small edits, and had a few comments.  

Final paragraph of the intro, first paragraph of discussion – The text discusses the use of data limited methods.  While we did use methods that are not as intensive as age structured assessments.  The methods we used are only possible in a data rich fishery with long fisheries independent time series.  Just do not want the reader to be looking to the study for examples of more traditional data limited methods like DCAC or LB-SPR only to find methods that needs lots of survey data.  Might also be good to be more explicit about saying most of the methods require fisheries independent data. 

Methods
Different operating models.  Just curious what the best way to describe the different operating models are. We used a single model (WHAM) to simulate everything, but parameterized it differently for different aspects (change in natural mortality, misrepresented catch, etc).  If someone said they used two different operating models, I would expect that they used two completely different models (e.g. WHAM and SS), not one model with different parameters for catch and mortality.  This may just be that I am not immersed enough in the MSE literature so my interpretation is not what others would think. 

Results
One area we discussed in the report was the difference in performance based on the cause of the retrospective pattern, M or catch.  Just checking that this was considered less important and not worth mentioning in the manuscript?  Based on the comments in the discussion, it seems like this came up.    I like Chris’s paragraph on the reference points and considering how much is going on in the study already (models, parameters, time periods, scenarios), I think clearly spelling out for the reader the challenge of the reference points is a good idea.  Not sure the M vs catch debate is actionable for a manager, they likely would not know which is happening, but it is interesting and seemed to come out in the plots.  If possible, it seems it would be good to include it

Does it matter that the results of the SCAA are not specifically mentioned till the discussion?

Discussion

Added a very short paragraph mentioning that many of the IBM’s had multiple options to make them more or less conservative in regards to SSB and F.  We selected one option, but the user could choose different risk tolerance/options. 

A big take away from the study was the SCAA.  Is it worth a paragraph near the end of the discussion mentioning any other big take aways: changes in selectivity were not really able to create the large Rho’s, for a bunch of methods OFL vs ABC was not a huge deal, etc?